agent/requirements.txt
livekit-agents[google]>=1.2.0,<2.0.0
python-dotenv==1.0.0
opencv-python-headless==4.8.1.78
numpy==1.26.2
pytesseract==0.3.10
Pillow==10.1.0
